+++ cw-vault-standard/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - `PreviewDeposit` query
+- `PreviewRedeem` query
 
 ## [0.3.3] - 2023-09-27